update edit trash

+++ laravel-app/app/Http/Controllers/Api/ProductController.php
@@ -15,6 +15,7 @@ use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Str;
 use Intervention\Image\ImageManagerStatic as Image;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\Validator;
 
 class ProductController extends Controller
 {
@@ -110,12 +111,12 @@ class ProductController extends Controller
 
                     // Lưu ảnh đầu tiên vào trường images của bảng products
                     if ($key == 0) {
-                        $product->images = $path;
+                        $product->image = $path;
                         $product->save();
                     } else {
                         // Lưu các ảnh còn lại vào bảng product_images
                         $productImage = new ProductImages();
-                        $productImage->product_id = $product->product_id;
+                        $productImage->product_id = $product->id;
                         $productImage->image = $path;
                         $productImage->author = $request->user()->name;
                         $productImage->status = $request['status'];
@@ -125,7 +126,7 @@ class ProductController extends Controller
 
                 // Lưu thông tin options của sản phẩm vào bảng Options
                 $options = new Options();
-                $options->product_id = $product->product_id;
+                $options->product_id = $product->id;
                 $options->color = $product->color;
                 $options->inch = $product->inch;
                 $options->author = $request->user()->name;
@@ -135,7 +136,7 @@ class ProductController extends Controller
                 // Lưu thông tin countdown của sản phẩm vào bảng CountDown (nếu có request)
                 if ($request->has('start_time') && $request->has('end_time')) {
                     $countdown = new CountDown();
-                    $countdown->product_id = $product->product_id;
+                    $countdown->product_id = $product->id;
                     $countdown->start_time = $request['start_time'];
                     $countdown->end_time = $request['end_time'];
                     $countdown->author = $request->user()->name;
@@ -148,7 +149,7 @@ class ProductController extends Controller
             return response()->json([
                 'status' => 'Created Successfully!, hihi',
                 'product' => $product
-            ], 201);
+            ], 200);
         }
     }
 
@@ -163,9 +164,48 @@ class ProductController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, Product $product)
+    public function edit($id)
     {
-        //
+        $data = Product::find($id);
+
+        if ($data) {
+            return response()->json([
+                'status' => 200,
+                'product' => $data,
+            ]);
+        } else {
+            return response()->json([
+                'status' => 404,
+                'message' => 'No Product Found',
+            ]);
+        }
+    }
+
+    public function update(Request $request, $id)
+    {
+        $validator = Validator::make($request->all(), [
+            'nameProduct' => 'required|string|max:255',
+            'category' => 'required',
+            'brand' => 'required',
+            'summary' => 'required|string|max:255',
+            'costProduct' => 'required|numeric',
+            'priceSale' => 'required|numeric',
+            'discount' => 'numeric',
+            'color' => 'required|string|max:255',
+            'inch' => 'required|numeric',
+            'start_time' => 'date',
+            'end_time' => 'date',
+            'detail' => 'required|string',
+            'status' => 'required',
+            'images.*' => 'image|mimes:jpeg,png,jpg,gif|max:2048',
+        ]);
+
+        if ($validator->fails()) {
+            return response()->json([
+                'status' => 422,
+                'errors' => $validator->messages(),
+            ]);
+        }
     }
 
     /**
@@ -179,35 +219,72 @@ class ProductController extends Controller
         $product->status = 0;
 
         if ($product->save()) {
-            $option = Options::where('product_id', $product->id)->update(['status' => 0]);
-            $countdown = CountDown::where('product_id', $product->id)->update(['status' => 0]);
-            $images = ProductImages::where('product_id', $product->id)->update(['status' => 0]);
-
+            Options::where('product_id', $product->id)->update(['status' => 0]);
+            CountDown::where('product_id', $product->id)->update(['status' => 0]);
+            ProductImages::where('product_id', $product->id)->update(['status' => 0]);
             return response()->json(['message' => 'Product has been softly deleted.']);
         } else {
             return response()->json(['message' => 'Failed to delete product.']);
         }
     }
+    public function trash()
+    {
+        $products = Product::onlyTrashed()->get();
+        return $products;
+    }
+
+    public function restore($id)
+    {
+        // tìm id có xóa tạm không
+        $product = Product::withTrashed()->find($id);
+
+        if (!$product) {
+            return response()->json(['message' => 'Product not found.'], 404);
+        }
+        // tìm id có trong danh sách xóa tạm không
+        if ($product->trashed()) {
+            $product->status = 1;
+            if ($product->save()) {
+                Options::where('product_id', $product->id)->update(['status' => 1]);
+                CountDown::where('product_id', $product->id)->update(['status' => 1]);
+                ProductImages::where('product_id', $product->id)->update(['status' => 1]);
+                $product->restore();
+            }
+            return response()->json(['message' => 'Product has been restored.', 'product' => $product->fresh()]);
+        } else {
+            return response()->json(['message' => 'Product is not deleted.']);
+        }
+    }
+
     /**
      * Remove
      */
-    public function remove(Product $product)
+    public function remove($id)
     {
-        $product->forceDelete();
-        return response()->json(['message' => 'Xóa vĩnh viễn thành công']);
+        // tìm id có xóa tạm không
+        $product = Product::withTrashed()->find($id);
+        $product_id = $product->id;
+
+        if (!$product) {
+            return response()->json(['message' => 'Product not found.'], 404);
+        }
+        // tìm id có trong danh sách xóa tạm không
+        if ($product->trashed()) {
+            // Xóa các bảng liên quan
+            Options::where('product_id', $product_id)->delete();
+            CountDown::where('product_id', $product_id)->delete();
+            ProductImages::where('product_id', $product_id)->delete();
+            $product->forceDelete();
+            return response()->json(['message' => 'Permanently deleted successfully']);
+        } else {
+            return response()->json(['message' => 'Product is not deleted.']);
+        }
     }
+
     /**
      * Trash
      */
-    public function trash()
-    {
-        return $products = Product::onlyTrashed()->paginate(25);
-    }
     /**
      * Restore
      */
-    public function restore($id)
-    {
-        return  Product::withTrashed()->find($id)->restore();
-    }
 }
